Moves SyntaxHighlight component to ui
